# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pakistan Accountability Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PublicDatabaseCategory =
  (typeof PublicDatabaseCategory)[keyof typeof PublicDatabaseCategory];

export const PublicDatabaseCategory = {
  electoral: "electoral",
  financial: "financial",
  corporate: "corporate",
  procurement: "procurement",
  payroll: "payroll",
  judicial: "judicial",
  land: "land",
  welfare: "welfare",
  foreign_assets: "foreign_assets",
  law_enforcement: "law_enforcement",
} as const;

export type PublicDatabaseAccessLevel =
  (typeof PublicDatabaseAccessLevel)[keyof typeof PublicDatabaseAccessLevel];

export const PublicDatabaseAccessLevel = {
  public: "public",
  partial: "partial",
  restricted: "restricted",
  unavailable: "unavailable",
} as const;

export type PublicDatabaseIntegrationStatus =
  (typeof PublicDatabaseIntegrationStatus)[keyof typeof PublicDatabaseIntegrationStatus];

export const PublicDatabaseIntegrationStatus = {
  integrated: "integrated",
  planned: "planned",
  manual: "manual",
  unavailable: "unavailable",
} as const;

export interface PublicDatabase {
  id: number;
  name: string;
  acronym: string;
  category: PublicDatabaseCategory;
  organization: string;
  description: string;
  dataTypes: string[];
  accessLevel: PublicDatabaseAccessLevel;
  url?: string | null;
  recordsEstimated: string;
  lastUpdated?: string;
  integrationStatus: PublicDatabaseIntegrationStatus;
  keyFields: string[];
  notes: string;
}

export type OfficialRiskTrend =
  (typeof OfficialRiskTrend)[keyof typeof OfficialRiskTrend];

export const OfficialRiskTrend = {
  rising: "rising",
  falling: "falling",
  stable: "stable",
} as const;

export interface Official {
  id: number;
  name: string;
  /** Partial CNIC for privacy (last 4 digits masked) */
  cnicPartial: string;
  position: string;
  party: string;
  province: string;
  constituency?: string;
  riskScore: number;
  riskTrend: OfficialRiskTrend;
  flagCount: number;
  assetsDeclared: string;
  lastUpdated: string;
}

export interface CompanyLink {
  companyName: string;
  secp_number: string;
  relationship: string;
  directorship: boolean;
  contractsValue: string;
}

export interface ContractLink {
  contractId: string;
  projectName: string;
  amount: string;
  awardedTo: string;
  connection: string;
  year: number;
  source: string;
}

export interface RelativeLink {
  name: string;
  relationship: string;
  companies: string[];
  contracts: number;
}

export interface FundingAllocation {
  district: string;
  amount: string;
  year: number;
  contractsBack: number;
  suspiciousLinks: boolean;
}

export type RiskFactorEvidenceLevel =
  (typeof RiskFactorEvidenceLevel)[keyof typeof RiskFactorEvidenceLevel];

export const RiskFactorEvidenceLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface RiskFactor {
  factor: string;
  weight: number;
  description: string;
  evidenceLevel: RiskFactorEvidenceLevel;
}

export type OfficialDetail = Official & {
  companies: CompanyLink[];
  contracts: ContractLink[];
  relatives: RelativeLink[];
  fundingAllocations: FundingAllocation[];
  riskFactors: RiskFactor[];
};

export type GraphNodeType = (typeof GraphNodeType)[keyof typeof GraphNodeType];

export const GraphNodeType = {
  official: "official",
  company: "company",
  relative: "relative",
  contract: "contract",
  district: "district",
  party: "party",
} as const;

export interface GraphNode {
  id: string;
  label: string;
  type: GraphNodeType;
  riskScore?: number | null;
  value?: string | null;
}

export interface GraphEdge {
  source: string;
  target: string;
  label: string;
  suspicious: boolean;
}

export interface ConnectionGraph {
  nodes: GraphNode[];
  edges: GraphEdge[];
}

export type RiskScoreCategory =
  (typeof RiskScoreCategory)[keyof typeof RiskScoreCategory];

export const RiskScoreCategory = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface RiskScore {
  officialId: number;
  officialName: string;
  position: string;
  province: string;
  score: number;
  category: RiskScoreCategory;
  topFlags: string[];
  lastCalculated: string;
}

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type AlertStatus = (typeof AlertStatus)[keyof typeof AlertStatus];

export const AlertStatus = {
  new: "new",
  reviewing: "reviewing",
  resolved: "resolved",
} as const;

export type AlertPatternType =
  (typeof AlertPatternType)[keyof typeof AlertPatternType];

export const AlertPatternType = {
  ghost_employee: "ghost_employee",
  circular_contract: "circular_contract",
  relative_enrichment: "relative_enrichment",
  fund_diversion: "fund_diversion",
  asset_mismatch: "asset_mismatch",
  procurement_anomaly: "procurement_anomaly",
} as const;

export interface Alert {
  id: number;
  title: string;
  description: string;
  severity: AlertSeverity;
  status: AlertStatus;
  officialId?: number | null;
  officialName?: string | null;
  detectedAt: string;
  patternType: AlertPatternType;
  estimatedValue?: string | null;
  dataSources: string[];
}

export interface SystemStats {
  totalOfficials: number;
  databasesConnected: number;
  alertsGenerated: number;
  highRiskOfficials: number;
  totalContractsScanned: number;
  estimatedFundsAtRisk: string;
  ghostEmployeesFound: number;
  lastFullScan: string;
}

export type ListOfficialsParams = {
  province?: string;
  party?: string;
  minRisk?: number;
  search?: string;
};

export type ListAlertsParams = {
  severity?: ListAlertsSeverity;
  status?: ListAlertsStatus;
};

export type ListAlertsSeverity =
  (typeof ListAlertsSeverity)[keyof typeof ListAlertsSeverity];

export const ListAlertsSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type ListAlertsStatus =
  (typeof ListAlertsStatus)[keyof typeof ListAlertsStatus];

export const ListAlertsStatus = {
  new: "new",
  reviewing: "reviewing",
  resolved: "resolved",
} as const;