# lib/api-zod/src/generated/types/connectionGraph.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pakistan Accountability Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
import type { GraphEdge } from "./graphEdge";
import type { GraphNode } from "./graphNode";

export interface ConnectionGraph {
  nodes: GraphNode[];
  edges: GraphEdge[];
}