# lib/api-zod/src/generated/types/graphEdge.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pakistan Accountability Intelligence System API
 * OpenAPI spec version: 0.1.0
 */

export interface GraphEdge {
  source: string;
  target: string;
  label: string;
  suspicious: boolean;
}