# lib/api-zod/src/generated/types/officialDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pakistan Accountability Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
import type { CompanyLink } from "./companyLink";
import type { ContractLink } from "./contractLink";
import type { FundingAllocation } from "./fundingAllocation";
import type { Official } from "./official";
import type { RelativeLink } from "./relativeLink";
import type { RiskFactor } from "./riskFactor";

export type OfficialDetail = Official & {
  companies: CompanyLink[];
  contracts: ContractLink[];
  relatives: RelativeLink[];
  fundingAllocations: FundingAllocation[];
  riskFactors: RiskFactor[];
};