# lib/api-zod/src/generated/types/publicDatabase.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pakistan Accountability Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
import type { PublicDatabaseAccessLevel } from "./publicDatabaseAccessLevel";
import type { PublicDatabaseCategory } from "./publicDatabaseCategory";
import type { PublicDatabaseIntegrationStatus } from "./publicDatabaseIntegrationStatus";

export interface PublicDatabase {
  id: number;
  name: string;
  acronym: string;
  category: PublicDatabaseCategory;
  organization: string;
  description: string;
  dataTypes: string[];
  accessLevel: PublicDatabaseAccessLevel;
  url?: string | null;
  recordsEstimated: string;
  lastUpdated?: string;
  integrationStatus: PublicDatabaseIntegrationStatus;
  keyFields: string[];
  notes: string;
}