# lib/api-zod/src/generated/types/relativeLink.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pakistan Accountability Intelligence System API
 * OpenAPI spec version: 0.1.0
 */

export interface RelativeLink {
  name: string;
  relationship: string;
  companies: string[];
  contracts: number;
}