# lib/api-zod/src/generated/types/riskScore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pakistan Accountability Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
import type { RiskScoreCategory } from "./riskScoreCategory";

export interface RiskScore {
  officialId: number;
  officialName: string;
  position: string;
  province: string;
  score: number;
  category: RiskScoreCategory;
  topFlags: string[];
  lastCalculated: string;
}